pyproject.toml
[project]
name = "fastapi-admin-backend"
version = "0.1.0"
description = "FastAPI backend for Vue3 admin UI following Agent-Oriented Architecture"
authors = [{ name = "Codex", email = "codex@example.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "uvloop>=0.19.0",
    "httptools>=0.6.0",
    "SQLAlchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "passlib[bcrypt]>=1.7.0",
    "python-jose[cryptography]>=3.3.0",
    "alembic>=1.13.0",
    "typer>=0.12.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0"
]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"
